LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/cache-storage-reporting-dedicated-worker.https-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: ReportingObserver

Harness Error (FAIL), message = ReferenceError: Can't find variable: ReportingObserver

TIMEOUT COEP support on DedicatedWorker. Test timed out
NOTRUN COEP-Report-Only support on DedicatedWorker. 
